pyproject.toml
[project]
name = "task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.4.0",
    "django-celery-beat>=2.7.0",
    "django-celery-results>=2.5.1",
    "django-cors-headers>=4.7.0",
    "django>=5.1.6",
    "djangorestframework-simplejwt>=5.4.0",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.8",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "redis>=5.2.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]
